[EGP] - Added 3DTracker object. - Added egpPos(N,V) (3D vector for use with the 3DTracker object)
This object is invisible and doesn't draw anything. You give it a 3D vector and it automatically runs ToScreen on it on the client. To use it, parent other objects to it. 

Can be used to make cross hairs or floating text etc. 

Only works on the EGP HUD. Behavior on the screen and emitter is undefined.