Change various e2 round functions to use garry's
Change the round functions for numbers, angles, vectors and vector2s to use garry's round function instead of computing it.
On my machine it's even faster this way, for Nebual it appears to be slower.

Either way, the way it was computed was erroneous and sometimes gave strange results.

Fixes #446