Made the tool skip a stage if there's only one output
Made the tool skip the output selection stage if there's only one output
on the target entity. The previous adv wire tool had this feature.

Also fixed a minor problem where the selected input would be set to 1
when you didn't want it to (such as after completing a wiring).