Added lua's os.time and os.date to E2
* Added `table date()` - Returns the server's current time and date as a
table
* Added `table date(time)` - Returns the specified time and date as a
table
* Added `number time()` - Returns the server's current time
* Added `number time(table data)` - Returns the specified time in
seconds
* Moved the old (now more or less obsolete) `number time(string
component)` function to the timer library.

I discovered that there was no way to convert a unix time to year,
month, week, hour, minute, etc etc without doing a whole lot of modulus
and division and crap like that. So I decided to add functions that
allow you to do this.