[Most likely won't work; I haven't that much experience] holoSubColor(N,V) - sets sub section color of a model like a player or a physgun.
I don't know if the function works, for I couldn't get wiremod to work successfully, and I don't have a bit of knowledge in Lua. I just constructed this with a tiny bit of TylerB's help, and I'll rely on that faith that it might work. If it doesn't, please modify the function in whatever way you want possible. Thanks.