Changed the method for scaling bones
Lines 196-200, the for loop has been reversed, starting from the back and working forward. The bone scale method prefers scaling in reverse order.

I am not sure about the efficiency of adding "self:SetModelScale((scale.x + scale.y + scale.z) / 3, 0)" directly after the bone scaling. When you set a hologram model to a ragdoll, it does not reach the "else self:SetModelScale((scale.x + scale.y + scale.z) / 3, 0) end" statement, and therefore only scales the bones and not the whole thing.

The issue that I have with the current method is that when you scale it normally, ie: holoScale( 1, vec(2,2,2) ), it does not only scale the entire model, but each bone as well. When this method is used, you have to loop through all of the bones if you want them reset to normal scale, after re-scaling the model it's self.