made E2s not save using workshop save
The server crashes whenever you save using the workshop save thingy when
there's an E2 spawned. This fixes that by not allowing E2s to be saved
at all.

I think the problem is somewhere in duplicator.CopyEnts (because it
copies every single value on the entire entity), but I could be wrong.