Input - allow external prop protection to dictact On/Off
#590 for Adv Input
#591 for Dual Input

I apologize if making 3 separate edits is redundant, I am not too familiar with Github yet.

------------------------------------------------------------------------------------------
There is currently no way to prevent players from messing with the owner's Adv Input without directly modifying this file.

Since the On/Off functions act like that of a Button, I chose PlayerUse as the hook. This will allow any custom or public prop protections to handle it instead. I believe Wiremod is using gamemode.Call for hooks included in the base gamemode, so I am using that instead of hook.Run.

There may be a better hook available for handling this, but I can't think of one at the moment. If this isn't merged, please at least look into adding a way to let external prop protections to handle this.