Made remote uploader display ops and cpu of all E2s (admin only)
This will be useful for admins, to see if an E2 is using a lot of cpu
time.
It is disabled by default, due to potentially using a lot of networking
(if there's a lot of E2s out). To enable it, go to the "Expression 2"
tab of the E2 editor options menu - there's a new checkbox there.