Added overlay to several entities which didn't have one
I found it annoying to have a bunch of entities on your contraption
(such as camera controller and eye pod) and have no way of knowing which
was which, because they're the same model. Changing model is one way of
knowing, I suppose, but this is neater.
Plus now the settings are displayed on cam controller and eye pod, so
other people looking at them know what's up.