Line 86 and the checkbox statement
The tool states:
"Determines wether the client side movement is local to the parent or not"
I don't actually see a reason why the angles should be parented too, it's supposed that only the camera position gets parented when it states movement. I'd rather suggest that there should be two options in the tool, one that allows you to parent the camera angles to the parent entity and another one that parents the camera position to the entity, so you will be able to have constant angles and parented position in the camera.