Added Wire Triggers and removed BIDIRARRAY uses
Wire Triggers are practically 3D Rangers. They output how many entities
are inside of them and an array of the entities inside. Players can
change the size, the offsets, and the conditions of these triggers.
However, I can't figure out how to rotate them with the receivers (that
output the trigger stuff), and believe me I've tried, so they're only
perpendicular to the world's axes.

BIDIRARRAY is only used in the 8x Table Duplexer and nowhere else. Not
only is it useless (at the moment), but it also takes out the usage of
using the duplexer with an ARRAY type.