Made DHDD more intuitive to use
AllowWrite now blocks any changes to the memory. Set AllowWrite to 0
just before duping the chip to ensure that the correct data is saved &
loaded properly.

In addition, unwiring the Data input no longer resets it. Use the Clear
input for that (Setting AllowWrite to 0 will not block clearing).

AllowWrite defaults to 1.