Extra entities to be filtered on cam controller
Hello there, I'm using the cam controller for a lot of things and I eventually create parented "contraptions" that have no actual constraints.
I'm not sure if the parented entities can be found through the constraints iteration (from the line 417 on my edit), I'm suggesting this edit so I'll be able to filter the entities that the Lua script is not able to find. I wasn't quite sure how the wire lib worked so I added two conditions on the line 421, to prevent lua errors.