Added emergency E2 shutdown
This feature is off by default, and will immediately halt the execution
and clear the memories of all E2s on the server if RAM usage spikes or
if the server's RAM usage becomes dangerously high.

I got the values by trial and error on a local server, so they might not
be optimal for larger dedicated servers, but that's why it's off by
default.