Added small additional description to applyForce
Yeah, I found some cases where e:applyForce was enabling collisions on the whole contraption it was constrained to, once I commented that line I didn't have problems with that again. I also know other people that had problems with applyForce and collisions aswell, and it turned out that entity:applyOffsetForce(v, entity:massCenter()) is a good replacement for entity:applyForce.

I probably should have suggested a change on the actual code than in the description, what should I do with this?