added check for the xmin being bigger than xmax
I think the check is important, since you can setup the screenscaling to be flippen horrizontallz or verticly so the xScale could be Vector2(100,100) and the yScale Vector2(0,0)

And I have a question. Should I still use `return { this.xScale[1], this.yScale[1], this.xScale[2], this.yScale[2] }` or should I add the action `GetScale`to return the scaling?