Make tps_pretty work with extended types
Previously, tps_pretty was called in some places with a table argument and
in some with a string. Calling it with a string would work okay unless it
contained an extended type (in which case the type eg. 'xv4' was seen as
three types).

tps_pretty now uses functionality from E2Lib to split the type ID strings
and get their human-readable names.

Fixes #1022.