Fix #519 - error if a function doesn't return
Unlike #1067 this performs checking for whether a function body returns
a value or not at runtime, and so it will only error when a function
does actually fail to return a value. This means it won't break
anybody's E2 scripts unless they were already currently broken (as
opposed to just hypothetically broken).