Floor address values for hi-speed read/write
This now sets up a consistent behaviour across all hi-speed entities:
the address given is floored in all situations. As well as being
consistent for the sake of being consistent, this also fixes #1065, in
which entities with bounded writable hi-speed memory could essentially
be used unbounded, as non-integer address could be passed in.