xwl:writeArray() - No longer null-terminate
Fixes #879

There is no need to automatically null terminate all arrays, readArray specifies the size of the area to read and manually adding a 0 to array in case it is needed for whatever reason is trivial. And there are seperate functions for writing/reading strings.

I added back the null-terminator for sub-tables (there probably is a better way to do that) since those would otherwise be hard to parse back into a array without mapping out all pointers beforehand.

An alternate solution would be a wirelink:writeArraySimple(address,array) function, that just writes the numerical contents of the array, without any sub-tables, null-terminations or other confusing stuff, which is certainly useful for many things including digital screens, where you just want to input a row of simple numbers.