Added wire friendslist
An entity that allows you to add SteamIDs to a list. When those players
connect to the server, the entity will output them in an array. Also
provides several different ways to check if a player is on the friends
list. Outputs 1 for 0.1 seconds when the player is on the list, -1 when
they are not, and 0 when the entity isn't checking anything.