Optimized digi screen
* Removed the uncompressed method of transfer (it wasn't even working).
All data is now always compressed before transfer.
* Made the processing of the data clientside happen more slowly, over
multiple frames. No more massive stutters when data is received (minor
stutters could still happen).
* Added an automatic buffering method to the serverside code. If a user
wants to use the digi screen in a way that causes many thousands of
writes per second (to load images or use ranger scanners for example),
it helps to buffer the changes and send it as large chunks. This feature
is activated automatically as needed.