Attempt to fix while perf with low quotas
When a server lowers their convars considerably, a while perf loop would
often hit tick quota, even though it should be protected against it by
this function. This attempts to fix that by adding another 200 ops
buffer in the perf function. This number is insignificant for servers
with the default convars (or higher), but it should greatly benefit
servers with lowered limits. the perf(n) function is unchanged, as it
allows you to more or less specify your own limit anyway.