Changed checkForce function
I changed this into a function to clamp the values instead of returning if they're inrange. I've went to each file containing the check function and replaced it with the clamp function. The only issue I've seen in the testing I've done is the direction for applyForce may be weird if the force is too large, but that should occur with the old function too since the max force hasn't been changed.