made most linkable entities in wire scan for vehicles
Made most entities which are linkable to vehicles scan for the closest
vehicle in the contraption if you link them to something which is not a
vehicle. Useful for scripted cars and also for contraptions where it's
hard to reach the vehicle inside.

Made the eyepod toolgun use the WireToolObj.SetupLinking feature.