Remove randomness from base_wire_entity RB update
Remove randomness from base_wire_entity RB update

If it's good enough to update randomly every 3 to 10 seconds, it's good
enough to update every 10 seconds. It'd still be better to have some
evidence of why this is needed, and I'd be happy for it to be removed in
future, but this at least makes things slightly less ugly and bizarre.

Fixes #581.