Run AdvDupe_FinishPasting for normal Duplicator
AdvDupe_FinishPasting is a hook that both advduplicator and advdupe2
call once the contraption has finished pasting. It's useful for E2s, as
running on "if (first() | duped())" means that it can run before your
contraption is only partially pasted. This means that for E2s that are
expecting to be pasted as part of a large contraption, it's only really
possible to use one of the non-standard duplicator tools, for
dupefinished()'s sake.

This patches the `duplicator` module to also call AdvDupe_FinishPasting
when it's done. It passes all the same data as advdupe2 does, which
means that it should be compatible with anything that's out there in the
wild.

The way the code is written also allows for the duplicator module
functions to start taking extra parameters and returning extra values
without breaking the code.

Fixes #1407.