Fix soundPlay with 0 duration never deregistering
Duration 0 is supposed to mean "play the entire sample" (unless fade is
also specified, in which case it really does mean 0). However, currently
we assume it plays forever, which means that if `soundPlay` is called
repeatedly with no `soundStop` or `soundPurge` then you very quickly hit
the `wire_expression2_maxsounds` limit.

Fixes #866.