Allow rebound key to trigger pod Light output
Fixes #1453. This adds a system in which the client tells the server
about its key bindings (for a limited subset of binds), and then the
server uses that to call hooks when specific binds are pressed. This
system looks a little overkill, but it'll likely be used by other parts
of Wiremod shortly.