Fix fun/sfun/mto instructions being opaque
There's a couple of places in the E2 compiler (Parser.DumpTree,
Optimizer.Process) where we walk over the E2 AST in a generic way: for
each for each AST node, it checks if any of its children are also AST
nodes and optimizes them. However, the fun, sfun, mto, kvtable and
kvarray nodes currently all store child nodes indirectly inside a table,
which means that the AST printer and the optimizer essentially can't see
any of their contents. This fixes that for fun, sfun, and mto.

This is a purely compile time change and won't have any effect on E2s at
runtime.