Simplify parsing and compiling of kvarray/kvtable
Instead of kvarray and kvtable being special parser cases, now all
function calls with key-value arguments generate a new kvcall
instruction (that replaces the kvarray and kvtable instructions).
The kvcall instruction still compiles to the same kvarray/kvtable
operators under-the-hood, but there's now less code duplication.

The new kvcall instruction fixes the issue with the kvarray and kvtable
instructions not being easily traversed as all other instructions are.
It also records the order that key-value pairs were specified in, which
is necessary for `array(1=a(), 2=b())` to have a predetermined order in
which `a` and `b` are called. (It doesn't fix this, but it's a necessary
first step.)