Refactor direct and indirect trigger queueing
'Trigger queueing' is when a variable that's declared as an @output
changes. When the E2 finishes executing, the value of that global
variable at the end is output. There are two cases:

1. Direct trigger queueing. This is the most simple case, and occurs
when a global @output variable is assigned a new value.
2. Indirect trigger queueing. This only occurs with mutable types -
tables and arrays. (gtables are also mutable types, but can't be output
over wires so don't do trigger queueing.) In this case, the value
assigned to the output stays the same, but the contents of that value
are changed. For example:

    @output T:table
    if (first()) {
        X = table() # direct trigger queued for X
    } else {
        local Y = X
	Y[1, number] = 1 # indirect trigger queued for all variables
	which refer to the same table as Y (which in this case is X)
    }

Previously this was handled by a more complicated system, implemented as
a callback in table.lua. This new system is simpler, and should be
faster as well.