Simplify var instructions at runtime
Performance-wise this will only be a small win, but it also cleans up
code - I'd like to remove all accesses to variables via self.Scopes, and
instead move towards everything knowing ahead of time (as in this
example) whether to use self.Scope or self.GlobalScope.

This also doesn't fix the "var" operator, which looks like it's only
ever used when the compiler attempts to transform the AST inline and
doesn't do a very good job of it.