Make DHDD's Size output deterministic
Currently, the DHDD size is computed with `#self.Memory`, and the
memory starts being entirely `nil` (and writing 0 to a memory cell
internally writes `nil` as a memory optimization). Given that the `#`
operator is nondeterministic when a table isn't a sequence, this makes
it essentially useless. Additionally, the DHDD memory is 0-indexed, but
Lua sequences are 1-indexed, which meant that writing to cell 0 wouldn't
change the size.

This changes the size output to a 'high water mark', by tracking the
largest address written to. The size now matches what's expected with
0-based indexing: if you write to cell 9, for example, then the size is
10.

Fixes #1581.