Fix text editor whole-word search with 1-char gap
In various places the text editor does things with whole words only:
it's possible to find and replace only whole words, and to count the
occurrences of a given whole world, and the editor also highlights every
occurrence of the selected whole word. The patterns we use to match word
boundaries are buggy: for example, at the moment higlighting and finding
will miss every other occurrence if there's a one-character gap between
the two.

This uses a consistent 'frontier pattern' throughout the code. This lets
is neatly match word boundaries without having to match beyond them.

Fixes #1601.