Fix infinite recursion when extending Wiremod ents
In our entities, we do:

    DEFINE_BASECLASS("base_wire_entity")

...which is replaced by Garry's Mod by:

    local BaseClass = baseclass.Get("base_wire_entity")

It also indirectly ends up setting `self.BaseClass` on the entity to its
parent. Throughout our code we ignore the mysterious `local BaseClass`
and instead use `self.BaseClass`. This works okay, until somebody uses
one of *our* classes as a base: imagine, for example, somebody extended
`gmod_wire_colorer`, which contains:

    function ENT:Think()
        self.BaseClass.Think(self)
	-- ...
    end

In the extended entity, `self.BaseClass` points to `gmod_wire_colorer`,
not to `base_wire_entity`, so this will be an infinite recursion.

The solution is just to use the file-local `BaseClass` variable that's
magically created for us.