Fix gates with nil-valued inputs
For some wire types, including ANY and WIRELINK, `nil` is a valid value.
When a gate uses one of these types, then at runtime things can break in
quite a subtle manner. If any value is `nil`, then the return value of
`ENT:GetActionInputs()` can be for example `{1, nil, 2}`. When this is
passed to `self.Action.output` with `unpack`, it may pass `1, nil, 2`
or it may just pass `1`. The gate then receives `nil` as the third
parameter when it was expecting a number, and things explode.