Optimize inside expressions we previously ignored
Previously, any AST node inside a call, methodcall, stringcall, kvtable,
kvarray, or switch instruction would be ignored by the optimizer. This
is because the optimizer assumed that all children of an AST node would
be represented as fields directly in that node, which isn't the case for
these instructions.

This teaches the optimizer to walk over these instruction nodes in a
specialized way which means we now fully reach into their contents and
optimize like any other node.