HoloClip Entity Reference
Lets you set which entity a local clip origin and direction are local to.

This is much better than doing the same with global clips, as there is no networking beyond the initial spawn. The result is a much more visibly smooth clip, even when the 'parent' entity (and thus the origin and direction) is moving and rotating.