Set holoClip local entity
Lets you set which entity a local clip origin and direction are local to.

This is much better than doing the same with global clips, as there is no networking beyond the initial spawn. The result is a much more visibly smooth clip, even when the 'parent' entity (and thus the origin and direction) is moving and rotating.

I've added an additional holoClip e2 function that takes a final entity argument, but, incidentally, the original two functions can be also fed an entity index as the 'isglobal' argument.