Fix swapping from/to RC circumventing driver check
Currently the RC only requires that one is linked to the pod (not driving) in order to call PlayerExited when you put the tool away. This allows the RC to pass the pod occupancy check when you pull the tool back out (PlayerExited sets HasPly to nil) and then it's possible to hijack a player's vehicle.

Simple change requires that you also be the driver of the pod in order to run PlayerExited.