Better outlines/line strips
Offloads the drawing of a string of lines, ie used on all outlines and egpLineStrip into a dedicated function.
This function attempts to decrease the gaps between line segments by streching the rectangular segments to line up at the outer corner of each bend for angles<=90 degrees.
That improves the quality of wide outline massively, but since drawTexturedRectRotated rounds the inputs, it still isn't perfect.