Remove a broken E2 peephole optimization
Previously, the E2 optimizer would replace `(mul 0 x)` with 0, for any
x. This also included when x is a vector, which means that the optimizer
produced incorrect results. The correct thing to do would be to replace
it with the zero value of the same type as the result of `(mul 0 x)`,
but this optimizer unfortunately runs before type-checking so that
information isn't available.

Fixes #1975.