added timer
Note:
Server now sends update time to client instead of client getting its own update time. this is intentional and prevents possible issues with a desynced curtime. I'm not sure if curtime is always synced in the source engine, but I lean toward "no" as the more likely answer and this is a very simple fix for that.

The initial sync (after beginning to look at an entity) is always done immediately to that specific player and is done using TCP (just like storm's code). After that, each later update while the player is looking at the entity is done every 0.1s and is done using UDP.

The "wire_overlay_txt" net message no longer exists because it serves no purpose.

The update timer is created and destroyed as necessary, as opposed to the existing code in the main wire repo where the timer is always running.