Use #2110's PlayerTable system
There are also other cores that need to use these new tables, I'm not sure if I want to add more to this PR though.

Thanks to gohidas pointing it out to me.. I got rid of the hook inside player.lua that cleaned up the player death / respawn events and replaced it with #2210's player table system.