Allow E2 to write (and read) files other than .txt
Allows E2 to write and read txt, dat json, jpg, xml, csv, and any other file extensions that are allowed by garrysmod. Useful if you for some unholy reason make a PNG writing library in E2 (I definitely didn't do that)

Adds a SHARED E2Lib function isValidFilePath that simply returns if it has an extension, and if that extension is on the whitelist.