Fixes, directive & global support
* Structs have default values so they can be used as global variables
* This also means they work in ``@persist``
* Edited the preprocessor to scan for structs in order to fix this
* Fixed errors in core/struct.lua
* Add ENT:GetE2Type which accesses the instance/compiler to get a type and falls back to wire_expression_types.
* Optimize print(...)
* Lower op cost of struct definition
* Remove struct operator

Structs are NOT ALLOWED IN I/O. This includes ``@outputs @inputs @trigger`` and gTables. This is because we can't easily compare whether two struct objects are the same, and it is already incredibly hacky and difficult to get the default value of a struct. I would need to start getting into messing with the WireLib / gates, which I really don't want to do.