Another yearly update
Haven't made any commits here as this is a particularly large change
* Renamed a lot of stuff from the old 'type' -> 'struct' that was left over from the first implementation
* Structs can be in I/O now, however you cannot have them as inputs. You can instead use the ``struct`` generic type. This acts basically like C's ``void`` type that it's a dangerous sort of untyped object
* Added xst:fields() which returns the fields of a struct as a table.
Note you cannot index a generic ``struct`` value, it will give you a compile time error. Maybe in the future this could be lifted or there could be type casting.
* Fixed xst == xst leaking a boolean value to E2
* Deprecate NORMAL type a little further in i/o
* Probably fix having more than 9 usertypes at once (Caught it before it was an issue)

Think this has gotten to more of a bugfixing area as now the type system is considerably less hacky and this works well enough in i/o.