Fix Wired CD Reporting Wrong Number of Tracks
The Wire CD reports the wrong number of usable tracks. At the moment, the medium CD model (models/venompapa/wirecd_medium.mdl) reports 8 tracks and 172 total sectors. There is no entry for track 8 in the entity's TrackSectors table which causes the CD Ray to give an erroneous sector index when attempting to read from it.

This change corrects this problem by reducing the total number of tracks on the disk by 1.