Add unit tests for E2
Adds basic unit tests that can be invoked by the server console using ``e2test``. This imitates an E2 chip virtually by using the world entity as it's owner, and as such doesn't act 1:1, but it should be good enough for most cases.

You won't be able to use this with the workshop wiremod which shouldn't matter as you should really just be doing unit test / debugging on the github version anyways.

Additionally added a lot more tests to the parsing "test" and fixed 1975 not compiling

All of this works on a new E2Lib function, ``runScript``, which runs an E2 script independent of an entity