Implement warning system for preprocessor/tokenizer/parser, plus more warnings (#2423)
* Add much more warnings & some overhauling

* Fix global var errors

Didn't convert certain parts of code to the new variable format.

Also fixes "unused variables" being reset whenever you assign to an ``@persist`` variable. Now it takes into account whether it is persistent and if you use it before the assignment, and will keep it marked as used rather than reset it.