Add variable and type autocomplete (#2497)
* Add variable and type autocomplete

Now not just directive variables, but any local or global variables in your script will autocomplete. Additionally, types will autocomplete after a colon for function/event definitions, and after the `function` keyword.

* Remove ast changes

Shouldn't be a part of this PR, and I don't want to potentially cause issues with the optimizer.