Rewrite extpp
500 sLOC => ~350 sLOC.

Mostly rewrote external preprocessor so it now supports outputting functions that use the non-legacy form of creating functions now.

My eventual plan is to deprecate the preprocessor as a whole but it's nice that having this allows for extensions to be ported over to the new system automatically.

Additionally rewrote e2functions using the varargs to use ...<name> instead.