Actually define the parameters
Got too used to using registerOperator which actually passes the arguments to the lua function directly rather than through a table as functions do.

I think it'd be possible to do that for functions but it'd require a lot of code duplication to be efficient. Maybe it could be looked into in the future.