Remove support for single quat/complex literals
You can no longer have `i`, `j` and `k` act as valid expressions. This is because it was causing a parsing error for methods.

This is a breaking change but I really doubt anyone used this behavior, and it is silly. Additionally I've had a warning for using them since I rewrote the tokenizer, so if anyone was (and still actively keeps up with E2), they would know to change to 1k/1j/1i.

This would fix #2544