Fix exit() and reset() not resetting scope
This wasn't a problem in the past, since statements would pcall their inner contents and pop their scopes before erroring. Now that we don't do that, since exit() and reset() just trigger an error, it caused the E2 to desync between runtime and compile time scope as it'd never pop its scope.

Now the functions reset the scope to the global scope before erroring.

This got the garfield pipe E2 working.