Enforce new void representation
Void type is represented as just nil now instead of an empty string.

To enforce this, check for empty string inside of compiling exprcall.

Ideally the api just wouldn't pass an empty string now but stuff like e2doc depends on that afaik. Just easier to go about it this way for now.