Fix keyvalue calls not accepting identifiers
This is an issue that's existed for 10+ years on the old compiler but nobody has found it since it's pretty niche to even use the key value table/array constructor functionality.

Hacky fix but couldn't think of any other way. This is the fault of two things: E2 not having braces like lua does for expressions in these kv constructors, and E2 using the assignment operator inside them as well. So not much we can do about it besides this.

Added test to ensure it works correctly.