Change perm checks and add weapon/ammo checks
Added check to stripWeapon to throw a runtime error in case the weapon class is invalid.

Changed logic for permissions to AND, so you could be either a friend or have cppi perms rather than needing both to access the functions.

Added check to the ammo functions to ensure the ammo type is valid, or else throw a runtime error

self.entity.player -> self.player

Changed invalid weapon error message for giveWeapon functions

Fix checkGive using incorrect variable (this) instead of (target)