Add default == operator in case it is absent
One goal of this PR is to minimize the amount of third party code needed in E2 extensions.

Now the equals operator can be omitted if the implementation intends to be a simple lua equals operation.

This not only saves on needing to implement a redundant function, relying on extension authors to maintain this, but leaves room for optimizations knowing that the operator is "simple", like compiling switch case to a lookup table of lambdas, which wouldn't be possible if the equals operator was overloaded since it'd use table indexing with the value.

Additionally added more test cases.

This behavior may also come to operator_is in the future. For now, since overloading it with valid nonsimple functions is common, it will stay necessary.