Rewrite compound operator, decrement/increment operators
This fixes the case of compound operators alongside increment/decrement not working with types that had legacy operators.

Now these will simply act as syntax sugar for assignment and use the operators internally, now that I've removed the compound assignment operators.

It simplifies the compiler implementation of these significantly and deduplicates a lot of code regarding variable usage