Enforce function returns at compile time (#2788)
* Enforce function returns at compile time

* Raise cost of dynamic functions

* Revert previous change

Forgot const optimization isn't currently being used. Needs an analyzer step.

Should really not have included that code in the rewrite but it's fine being there for now.

* Add test cases

* Simplify

No need for `returned` field, can just check if scope is dead.

* Add switch case return logic

Adds logic for switch case to be the last statement in a function and be detected for return values.